Created a commands folder
Put commands in a file and in folder. It will make it easier to look at
and sort through later. The prefix right now is "!".
Some things like "frazee" aren't really commands and just happen so we
can still keep that where it is.

I commented a lot to try and help.